Refactor GenericRightPaneComponent to accept a ReactComponent as its content
